requirements.txt
fastapi==0.127.1
joblib==1.5.3
python-jose==3.3.0
pandas==2.3.3
prometheus_fastapi_instrumentator==7.1.0
pydantic==2.12.5
python-dotenv==1.2.1
redis==7.1.0
scikit_learn==1.8.0
starlette==0.50.0
uvicorn==0.38.0
